Revert "Add convex hull as an op. Switch from doubles to floats for convex hull."
This reverts commit db9534940daf2f2c93dcb7f60b77f8a3f7656236.

+++ com.ixxy.polyhydra/QuickHull3D/Runtime/Point3d.cs
@@ -45,6 +45,6 @@ namespace QuickHull3D
 	    /// <param name="x">first element</param>
 	    /// <param name="y">second element</param>
 	    /// <param name="z">third element</param>
-        public Point3d(float x, float y, float z) : base(x, y, z) { }
+        public Point3d(double x, double y, double z) : base(x, y, z) { }
     }
 }
